+++ src/dom-utils.ts
@@ -38,3 +38,25 @@ export function distributeSpans(text: string): string[] {
   if (liveSpans.length) throw "Unbalanced <span>s in " + text;
   return outLines;
 };
+
+/**
+ * Adds soft wrap markers between all characters in a DOM element.
+ */
+export function addSoftBreaks(el: HTMLElement) {
+  var softBreak = '\u200B';
+  walkTheDOM(el, function(node) {
+    if (node.nodeType !== 3) return;
+    var text = (node as Text).data;
+    text = text.split('').join(softBreak);
+    node.nodeValue = text;
+  });
+}
+
+function walkTheDOM(node: Node, func: (n: Node) => void) {
+  func(node);
+  let n = node.firstChild;
+  while (n) {
+    walkTheDOM(n, func);
+    n = n.nextSibling;
+  }
+}
